Move system from plain SQL queries to Spring Data JPA
This should make future extensions to the system much easier to develop. It is no longer necessary to write simple SQL queries over and over again as Spring now handles this for us.
All CRUD actions with the database (which most extensions do) are now extremely simple to add and adjust in case of future changes.
In addition to making extensions easier to implement, the system is now also easier to maintain as the code has been simplified and shortened.